# perf: optimize build options + improve to_camel() performance