--- a/Cargo.toml
+++ b/Cargo.toml
@@ -9,4 +9,4 @@ name = "casers"
 crate-type = ["cdylib"]
 
 [dependencies]
-pyo3 = "0.22.3"
+pyo3 = { version = "0.23.4", features = ["extension-module"] }