@daxartio 2025-02-03
Thank you for the significant improvement to this part of the code.