@deedy5 2025-02-03
Есть еще 2 момента:
1) функции работают только с ascii, но можно исправить;
2) если размер инпута известен заранее и не превышает конкретного значения, то можно сделать еще намного быстрее используя буфер на стэке.