@daxartio 2025-02-03
1. Да, в этом есть проблема. Так как было сделано для pydantic моделей, казалось ascii достаточно в реальных кейсах. Будет круто если есть желание добавить функционал.
2. Можно попробровать заранее выделить 2^n памяти на стеке под буфер. В реальной жизин это покроет 99%. Тоже будет супер.